+++ js/const.js
@@ -1,0 +1,19 @@
+// constants
+const TYPES_OF_APARTMENT = ['palace', 'flat', 'house', 'bungalow'];
+const CHECKOUT_TIMES = ['12:00', '13:00', '14:00'];
+const FACILITIES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const SOURCES_OF_PHOTOS = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+const QUANTITY_OF_RANDOM_ANNOUNCEMENT = 10;
+const FIRST_VALUE_OF_LATITUDE = 35.65000;
+const SECOND_VALUE_OF_LATITUDE =  35.70000;
+const FIRST_VALUE_OF_LONGITUDE = 139.70000;
+const SECOND_VALUE_OF_LONGITUDE = 139.80000;
+const NUMBER_OF_COORDINATE_POINTS = 5;
+const MIN_RANDOM_PRICE = 2500;
+const MAX_RANDOM_PRICE = 15000;
+const MAX_GUESTS_NUMBER = 16;
+const MAX_ROOMS_NUMBERS = 8;
+const MAX_NUMBER_FOR_AVATAR = 8;
+
+export {TYPES_OF_APARTMENT, CHECKOUT_TIMES, FACILITIES, SOURCES_OF_PHOTOS, QUANTITY_OF_RANDOM_ANNOUNCEMENT, FIRST_VALUE_OF_LATITUDE, SECOND_VALUE_OF_LATITUDE,
+  FIRST_VALUE_OF_LONGITUDE, SECOND_VALUE_OF_LONGITUDE, NUMBER_OF_COORDINATE_POINTS, MIN_RANDOM_PRICE, MAX_RANDOM_PRICE, MAX_GUESTS_NUMBER, MAX_ROOMS_NUMBERS, MAX_NUMBER_FOR_AVATAR};
